# docs/openapi_v9.yaml
openapi: 3.0.3
info:
  title: Plan Charge v9 API
  version: 1.0.0
  description: |
    OpenAPI spec for Plan Charge v9 backend (FastAPI) covering identity, planning, integrations, and reporting.
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    FromDate:
      name: from
      in: query
      schema: { type: string, format: date }
    ToDate:
      name: to
      in: query
      schema: { type: string, format: date }
  schemas:
...
    Absence:
      type: object
      required: [person_id, start_date, end_date, type]
      properties:
        id: { type: string, format: uuid }
        person_id: { type: string, format: uuid }
        start_date: { type: string, format: date }
        end_date: { type: string, format: date }
        type: { type: string }
        hours_per_day: { type: number, nullable: true }
        notes: { type: string, nullable: true }
...
paths:
...
  /absences:
    get:
      tags: [Calendars]
      summary: List absences
      parameters:
        - name: person_id
          in: query
          schema: { type: string, format: uuid }
        - { $ref: '#/components/parameters/FromDate' }
        - { $ref: '#/components/parameters/ToDate' }
      responses:
        '200': { description: OK }
    post:
      tags: [Calendars]
      summary: Create absence
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Absence' }
      responses:
        '201': { description: Created }
...